Fix monoid.numpy.fmax identity for INT16/INT32/INT64 (mapnumpy=False)
With config["mapnumpy"]=False, monoid.numpy.fmax for INT16/INT32/INT64
used int8's minimum (-128) as its identity instead of the type's own
minimum. Because SuiteSparse seeds a non-empty max-style reduction with
the monoid identity, an fmax reduction over values below -128 returned
-128 instead of the true maximum. Example (mapnumpy=False): fmax-reduce
of [-1000, -2000, -3000] over INT16 returned -128; it now returns -1000.

The identities now match the builtin maximum monoid byte-for-byte. INT8
(-128), the unsigned types (0), and the entire fmin block are unchanged;
_fmin_is_float (imported by semiring/numpy.py) is untouched.

Pre-existing since a22be711 (Oct 2022); not a regression, and only
reachable with mapnumpy=False (with mapnumpy=True fmax resolves to the
builtin max monoid). Add a regression test that reproduces the bug
end-to-end, guarded to skip (not error) on the numba #8478 combo where
the integer identities are absent.

+++ graphblas/monoid/numpy.py
@@ -103,11 +103,11 @@ if (
             "BOOL": False,
             "INT8": _np.iinfo(_np.int8).min,
             "UINT8": 0,
-            "INT16": _np.iinfo(_np.int8).min,
+            "INT16": _np.iinfo(_np.int16).min,
             "UINT16": 0,
-            "INT32": _np.iinfo(_np.int8).min,
+            "INT32": _np.iinfo(_np.int32).min,
             "UINT32": 0,
-            "INT64": _np.iinfo(_np.int8).min,
+            "INT64": _np.iinfo(_np.int64).min,
             "UINT64": 0,
         }
     )
